+++ airbyte-integrations/connectors/source-mailchimp/source_mailchimp/source.py
@@ -13,7 +13,7 @@ from airbyte_cdk.sources.streams import Stream
 from airbyte_cdk.sources.streams.http.auth import TokenAuthenticator
 from requests.auth import AuthBase
 
-from .streams import Automations, Campaigns, EmailActivity, Lists, Reports
+from .streams import Automations, Campaigns, EmailActivity, Lists, Reports, Unsubscribes
 
 
 class MailChimpAuthenticator:
@@ -70,4 +70,5 @@ class SourceMailchimp(AbstractSource):
             Automations(authenticator=authenticator),
             EmailActivity(authenticator=authenticator, campaign_id=campaign_id),
             Reports(authenticator=authenticator),
+            Unsubscribes(authenticator=authenticator, campaign_id=campaign_id),
         ]