+++ airbyte-cdk/python/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.52.4
+File CDK: Add unstructured parser
+
+## 0.52.3
+Update source-declarative-manifest base image to update Linux alpine and Python
+
+## 0.52.2
+
+
+## 0.52.1
+Add max time for backoff handler
+
 ## 0.52.0
 File CDK: Add CustomFileBasedException for custom errors